dinoscholar/requirements.txt
flask>=3.0,<4.0
flask-sqlalchemy>=3.1,<4.0
sqlalchemy>=2.0,<3.0
click>=8.1,<9.0
rich>=13.0,<14.0
markdown>=3.5,<4.0
python-frontmatter>=1.1,<2.0
pytest>=8.0,<9.0
pytest-cov>=5.0,<6.0
ruff>=0.3,<1.0
